+++ avaliacoes/pets/src/main/java/controller/VeterinarioController.java
@@ -1,0 +1,5 @@
+package controller;
+
+public class VeterinarioController {
+    
+}
